+++ TelCo.ColorCoder/ColorDefition.cs
@@ -1,0 +1,18 @@
+using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Text;
+
+namespace TelCo.ColorCoder.SubModule
+{
+    class ColorDefination
+    {
+        public static readonly Color[] colorMapMajor;
+        public static readonly Color[] colorMapMinor;
+        static ColorDefination()
+        {
+            colorMapMajor = new Color[] { Color.White, Color.Red, Color.Black, Color.Yellow, Color.Violet };
+            colorMapMinor = new Color[] { Color.Blue, Color.Orange, Color.Green, Color.Brown, Color.SlateGray };
+        }
+    }
+}
